Refactor: Move TV show details entities to separate files
