+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/CastEntity.kt
@@ -5,31 +5,3 @@ data class CastEntity(
     val crew: List<CrewMemberEntity>,
     val id: Int
 )
-
-data class CastMemberEntity(
-    val adult: Boolean,
-    val gender: Int,
-    val id: Int,
-    val knownForDepartment: String,
-    val name: String,
-    val originalName: String,
-    val popularity: Double,
-    val profilePath: String?,
-    val character: String,
-    val creditId: String,
-    val order: Int
-)
-
-data class CrewMemberEntity(
-    val adult: Boolean,
-    val gender: Int,
-    val id: Int,
-    val knownForDepartment: String,
-    val name: String,
-    val originalName: String,
-    val popularity: Double,
-    val profilePath: String?,
-    val creditId: String,
-    val department: String,
-    val job: String
-)
