+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/CrewMemberEntity.kt
@@ -1,0 +1,15 @@
+package com.london.domain.entity.tvshowdetails
+
+data class CrewMemberEntity(
+    val adult: Boolean,
+    val gender: Int,
+    val id: Int,
+    val knownForDepartment: String,
+    val name: String,
+    val originalName: String,
+    val popularity: Double,
+    val profilePath: String?,
+    val creditId: String,
+    val department: String,
+    val job: String
+)
