+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/ImageItemEntity.kt
@@ -1,12 +1,5 @@
 package com.london.domain.entity.tvshowdetails
 
-data class TvShowImagesEntity(
-    val backdrops: List<ImageItemEntity>,
-    val id: Int,
-    val logos: List<ImageItemEntity>,
-    val posters: List<ImageItemEntity>
-)
-
 data class ImageItemEntity(
     val aspectRatio: Double,
     val height: Int,
