+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowCreatorEntity.kt
@@ -1,0 +1,10 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowCreatorEntity(
+    val id: Int,
+    val creditId: String,
+    val name: String,
+    val originalName: String,
+    val gender: Int,
+    val profilePath: String?
+)
