+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowDetailsEntity.kt
@@ -34,69 +34,3 @@ data class TvShowDetailsEntity(
     val voteAverage: Double,
     val voteCount: Int
 )
-
-data class TvShowCreatorEntity(
-    val id: Int,
-    val creditId: String,
-    val name: String,
-    val originalName: String,
-    val gender: Int,
-    val profilePath: String?
-)
-
-data class TvShowGenreEntity(
-    val id: Int,
-    val name: String
-)
-
-data class TvShowEpisodeEntity(
-    val id: Int,
-    val name: String,
-    val overview: String,
-    val voteAverage: Double,
-    val voteCount: Int,
-    val airDate: String,
-    val episodeNumber: Int,
-    val episodeType: String,
-    val productionCode: String,
-    val runtime: Int?,
-    val seasonNumber: Int,
-    val showId: Int,
-    val stillPath: String?
-)
-
-data class TvShowNetworkEntity(
-    val id: Int,
-    val logoPath: String?,
-    val name: String,
-    val originCountry: String
-)
-
-data class TvShowProductionCompanyEntity(
-    val id: Int,
-    val logoPath: String?,
-    val name: String,
-    val originCountry: String
-)
-
-data class TvShowProductionCountryEntity(
-    val iso31661: String,
-    val name: String
-)
-
-data class TvShowSeasonEntity(
-    val airDate: String?,
-    val episodeCount: Int,
-    val id: Int,
-    val name: String,
-    val overview: String,
-    val posterPath: String?,
-    val seasonNumber: Int,
-    val voteAverage: Double
-)
-
-data class TvShowSpokenLanguageEntity(
-    val englishName: String,
-    val iso6391: String,
-    val name: String
-)
