+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowGenreEntity.kt
@@ -1,0 +1,6 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowGenreEntity(
+    val id: Int,
+    val name: String
+)
