+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowImagesEntity.kt
@@ -1,0 +1,8 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowImagesEntity(
+    val backdrops: List<ImageItemEntity>,
+    val id: Int,
+    val logos: List<ImageItemEntity>,
+    val posters: List<ImageItemEntity>
+)
