+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowNetworkEntity.kt
@@ -1,0 +1,8 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowNetworkEntity(
+    val id: Int,
+    val logoPath: String?,
+    val name: String,
+    val originCountry: String
+)
