+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowProductionCompanyEntity.kt
@@ -1,0 +1,8 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowProductionCompanyEntity(
+    val id: Int,
+    val logoPath: String?,
+    val name: String,
+    val originCountry: String
+)
