+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowProductionCountryEntity.kt
@@ -1,0 +1,6 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowProductionCountryEntity(
+    val iso31661: String,
+    val name: String
+)
