+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowSeasonEntity.kt
@@ -1,0 +1,12 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowSeasonEntity(
+    val airDate: String?,
+    val episodeCount: Int,
+    val id: Int,
+    val name: String,
+    val overview: String,
+    val posterPath: String?,
+    val seasonNumber: Int,
+    val voteAverage: Double
+)
