+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowSpokenLanguageEntity.kt
@@ -1,0 +1,7 @@
+package com.london.domain.entity.tvshowdetails
+
+data class TvShowSpokenLanguageEntity(
+    val englishName: String,
+    val iso6391: String,
+    val name: String
+)
